Split body_capture_mac.py into class files
