+++ Constants.py
@@ -1,0 +1,9 @@
+# Constants
+CAPTURE_FPS = 60  # Camera capture rate
+DEFAULT_VIDEO_FPS = 25  # Default video recording FPS
+TEXTURE_WIDTH = 1280
+TEXTURE_HEIGHT = 720
+VIDEO_CODEC = 'mp4v'
+VIDEO_FORMAT = '.mp4'
+BUFFER_SIZE = 30  # Frame buffer size
+CSV_BUFFER_SIZE = 100  # CSV data buffer size
